requirements.txt
flask==2.3.3
spacy==3.7.2
PyPDF2==3.0.1
python-docx==0.8.11
scikit-learn==1.3.0
numpy==1.25.2
werkzeug==2.3.7
sentence-transformers
pandas